feat: Lazily load 2qb gate (#505)
* feat: Lazily load 2qb gate

Rather than fetching the 2qb gate of devices on backend construction,
this commit changes the behavior of the backend to only fetch devices
once the 2qb gate is requested and caches the value as a cached_property
using a decorator. As a side effect it removes the need for the user to
log in to construct a backend.

Changes to tests that were required due to the backend now requiring
credentials to construct have been reverted.